WIP: Remove toolbar jitter
This tries to address jitter of the toolbar (especially noticeable when
scrolling horizontally) by using position: sticky and adding a small margin
to the bottom of the browser frame.

This works, but this small margin also means the page doesn't appear to
scroll at all while in the margin, which is a bit less desirable. Ideally,
the page would continue to scroll and this margin could then be much
bigger, but maybe this is very slightly better than it is at the moment?