WIP: Remove toolbar jitter using timeout
Try to remove the toolbar jittering on/off by introducing a timeout before
changing the state (except when at the extents, in which case, the toolbar
shows/hides immediately).

This works well, but unfortunately obviously shows the synchronously
updating top of the browser frame... Possibly a bug that this top part
doesn't scroll smoothly with the rest of the content?